requirements.txt
flask==3.0.3
python-dotenv==1.0.1
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
